Restructure of file organization
