+++ .gitignore
@@ -1,0 +1,2 @@
+/Automated_Media_Ingest.exe
+.idea
